+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+- Check replacing indexes
+
 ## 0.1.0 (2022-01-17)
 
 - First release